working with vectors and pointers

+++ ejercicio6/functions.h
@@ -1,8 +1,8 @@
 #ifndef FUNCTIONS_H_INCLUDED
 #define FUNCTIONS_H_INCLUDED
 
-double obtenerSeno(int, int);
-double potencia(int, int);
+double obtenerSeno(float, float);
+double potencia(float, int);
 double factorial(int);
 
 #endif // FUNCTIONS_H_INCLUDED
